packages/cli/templates/voice/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openai==1.6.1
anthropic==0.8.1
httpx==0.25.2
pyyaml==6.0.1
pydantic==2.5.2
pydantic-settings==2.1.0
twilio==8.10.0
python-multipart==0.0.6
assemblyai>=0.35.0
